API reference

ASEconvert.ase — Constant

Global constant representing the ase python module available from Julia.

ASEconvert.convert_ase — Method

convert_ase(system::AbstractSystem)

Convert a passed system (which satisfies the AtomsBase.AbstractSystem interface) to an ase.Atoms datastructure. Conversions to other ASE objects from equivalent Julia objects may be added in the future.